Instance-level security is not implemented
Problem
-------

The DDS security specification specifies functions for checking that
local or remote participant can register or dispose an instance.
Currently, these methods are defined in the appropriate interfaces but
they are not called.

To complicate matters, the functions are defined using DynamicData.
Thus, a prerequisite is to build up support for DynamicData where a
custom plugin for instance-level security is viable.

Solution
--------

* Build up support for DynamicData.

* Add calls to check when registering/disposing instances.